ml-lens/evals/requirements.txt
openai>=1.30.0
python-dotenv>=1.0.0
torch>=2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
